+++ src/test/java/no/entur/uttu/UttuTestApp.java
@@ -15,7 +15,7 @@
 
 package no.entur.uttu;
 
-import no.entur.uttu.organisation.legacy.EnturLegacyOrganisationRegistry;
+import no.entur.uttu.ext.entur.organisation.EnturLegacyOrganisationRegistry;
 import no.entur.uttu.repository.generic.ProviderEntityRepositoryImpl;
 import no.entur.uttu.security.UttuSecurityConfiguration;
 import no.entur.uttu.stopplace.DefaultStopPlaceRegistry;
